pyproject.toml
[project]
name = "wangwang-rpa"
version = "0.1.0"
description = "旺旺RPA自动化系统 - 自动化处理旺旺即时通讯消息的RPA工具"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pyyaml>=6.0",
    "selenium>=4.0",
    "webdriver-manager>=4.0",
]
[project.optional-dependencies]
dev = [
    "black>=24",
    "isort>=5",
    "mypy>=1.12",
    "ruff>=0.7",
    "pytest>=8",
    "pytest-cov>=5",
]
[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-ra -q --cov=src --cov-report=term-missing"

[tool.coverage.run]
source = ["src"]
